--- .env.production
@@ -1,1 +1,0 @@
-GENERATE_SOURCEMAP=false
