--- src/types/i18n.d.ts
@@ -1,4 +1,0 @@
-declare module "@dhis2/d2-i18n" {
-    export function t(value: string, namespace?: object): string;
-    export function changeLanguage(locale: string);
-}
